+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Add an Affiliation model and associates it to the Person abstract model through a ManyToManyField
 * Update the ``import_journals_from_fedora`` management command to import affiliations
+* Add new model admin classes
 
 0.2.0
 -----